+++ R/zzz.R
@@ -5,6 +5,7 @@
 
   toset <- !(names(opts.LandR) %in% names(opts))
   if (any(toset)) options(opts.LandR[toset])
+  .setScanfiMirror()
 
   invisible()
 }